Implement replicateU in terms of loopU (comments below)
This patch defines replicateU as a loopU (essentially a mapU) over unitU
and nukes the loopU/replicateU rewrite rule. This seems to work quite well
but has revealed an interesting problem: all of a sudden,

  map_replicate f n = mapU f . replicateU n

didn't get fused any more.

The reason for this seems to be quite subtle. Fusion didn't happen here
because replicateU didn't get inlined, apparently because GHC can't figure out
that loopU is strict in the array. After inling mapU we get

  map_replicate f n x = loopU (...) (replicateU n x)

loopU isn't considered an interesting continuation and hence replicateU isn't
inlined despite the INLINE pragma. I do have to admit that I don't quite
understand why both mapUs get inlined in

  map_map f g = mapU f . mapU g

I *think* this is because mapU has 4 arguments (including 2 dictionaries)
whereas replicateU only has 3 and GHC seems to take the number of arguments
into account when deciding whether to inline things.

Anyway, this is of course easily cured by adding a seq to the definition of
loopU which this patch does as well. The rudimentary tests in test/fusion all
pass now. However, now a loopU/loopU fusion doesn't happen in the primes
benchmark. This is somehow caused by the seq and definitely not by the new
replicateU (it works without the seq). I haven't figured out why this is -
the function is just too big. The benchmark doesn't seem to run slower because
of that, though.

The above does raise an interesting question, though. GHC can't possibly figure
out that loopU is strict in the array - loopU is strict in things like
(lengthU arr) but lengthU is a class method and GHC has no way of knowing that
lengthU is always strict; nor do we have a (direct) way of telling it.
Therefore, GHC doesn't infer any useful strictness information for many of
our combinators which is bound to interfere with optimisations. What do we do
about it?