Add a fuseable version of lengthU (hack)
This is a quick hacks which adds lengthU' to Flat/Sums; since it is
implemented as a fold, it can be fused with loopU. Counting sort runs twice as
fast if it is used in 'position'.

This will go away once we decide on how to make lengthU itself fuseable, but
I'd like to have this for testing for now.