Add DT instance for MaybeS a
This assumes that any MDT can be frozen even if not all of its elements
have been initialised. This works at the moment, but is perhaps somewhat
ugly. We do need distributed Maybes for folds and scans over distributed
arrays, though (I think, at least).

This can be easily generalised to distributed sums but I'm not sure if
it's worth it.