Remove SUArr
This is a very big patch which removes all traces (hopefully) of the SUArr
data type. The new convention is to pass segment descriptors and data arrays
separately. A lot of segmented operations have been deleted since they don't
make sense any more. Some operations (such as packCU/packCUP) do make sense
but should be implemented in terms of the abstract backend interface, much
like repeat_c is now.

At the moment, fusion for segmented operations is rather limited which is not
a big change from before. Also, the lifted library must be reorganised a
little to take full advantage of the new backend interface. Some examples have
not been adapted to the new interface yet.