Add experimental support for dtrace-based profiling
This is rather hackish at the moment but here is how it works (on OS X only).
We manually build libdph-trace.dylib, manually edit Extra-Lib-Dirs in
dph-base.cabal to point to it, manually turn on the DTrace flag and rebuild
DPH from scratch (as in rm */dist-install; make stage=2). Then, we build an
example and run it as follows:

  sudo dtrace -c './seq/quickhull 100000' -s loop-profile.d

This should produce a table which shows how much time is spent in which loops.