Trace all stream-based loops
We extend the stream data type with an additional String which says how the
stream was produced:

  data Stream a = forall s. Stream ... String

Now, whenever we create or transform a stream we record the function that did
it:

  replicateS n x = Stream ... (sNoArgs "replicateS")
  mapS f (Stream ... c) = Stream ... ("mapS" `sArgs` c)

Real loops that actually consume streams pass this string to dtrace. The
somewhat funny sNoArgs/sArgs setup is because I want to use ByteStrings
instead of Strings eventually and hope that GHC will be able to just create
big string literals instead of concatenating lots of small ones.