Completely change how mapD and friends are implemented
The basic combinators are now

generateD :: DT a => Gang -> (Int -> a) -> Dist a
imapD :: (DT a, DT b) => Gang -> (Int -> a -> b) -> Dist a -> Dist b

The Int argument is the index of the current thread. There is also

generateD_cheap :: DT a => Gang -> (Int -> a) -> Dist a

which is equivalent to generateD but just runs the computation sequentially.