Replace strict pairs by normal lifted pairs in distributed types
We use (a,b) instead of (a :*: b) in distributed types now. We will get rid of
(a :*: b) altogether once we move to vector so this is a change we'd have to
make anyway but happily, it also helps with performance a bit. In particular,
it allows us to trigger LiberateCase much less frequently.