Add Makefile code for building Cabal packages.
To make the Cabal packages for DPH, in a GHC build tree:

 $ cd libraries/dph
 $ make -f sdist.mk

This makes the packages and prints a list of the package files.

You can also do it outside of a GHC build tree, but tweak the GHC
setting in sdist.mk first (or pass a value for GHC on the make command
line).

Then upload each package to Hackage, e.g.:

 $ cabal upload dp-base/dist/dph-base-0.5.tar.gz