Reorganise the way the lifted reference implementation works
The lifted reference implementation is now in dph-lifted-base. The PArray functions there have no PA dictionaries. We need this for comparing against other PArray functions, but it means this module can't be used with the vectoriser by itself. The dph-lifted-boxed package will contain the functions that take PA dictionaries, and can be used with the vectoriser.